+++ app/models/interfaces/Route.ts
@@ -1,5 +1,6 @@
 export default interface Route {
   name: string;
   title: string;
-  component: React.FC<any>
+  component: React.FC<any>;
+  headerRight?: () => JSX.Element;
 }