+++ app/screens/balance/components/mainMenu/index.tsx
@@ -25,7 +25,6 @@ const MainMenu: React.FC<MainMenuProps> = ({
   <ScrollView showsHorizontalScrollIndicator={false} horizontal={true} style={styles.flatListMenu}>
     <ImageButton buttonTitle='List Transaction' imageBase64={listMenu[0].data} onPress={() => navigation.navigate('Transactions')} />
     <ImageButton buttonTitle='Balance History' imageBase64={listMenu[3].data} onPress={() => navigation.navigate('BalanceHistory')} />
-    <ImageButton buttonTitle='Search Transaction' imageBase64={listMenu[9].data} onPress={() => navigation.navigate('SearchTransaction')} />
     <ImageButton buttonTitle='Add Transaction' imageBase64={listMenu[2].data} onPress={() => navigation.navigate('CreateTransaction', {})} />
     <ImageButton buttonTitle='Add Category' imageBase64={listMenu[1].data} onPress={() => navigation.navigate('CreateCategory', {})} />
     <ImageButton buttonTitle='Renew Balance' imageBase64={listMenu[4].data} onPress={onRenewButtonPress} />