+++ app/screens/transactions/index.tsx
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from "react"
 import { SectionList, TouchableOpacity } from "react-native"
+import { SearchIcon } from "native-base";
 import { NativeStackScreenProps } from "@react-navigation/native-stack"
 
 import Transaction from "../../models/schemas/TransactionSchema";
 import RootStackParamList from "../../models/interfaces/RootScreensParams"
 
-import { BottomButton, ItemFlatList, SectionHeader } from "../../components";
+import { BottomButton, ItemFlatList, SearchInput, SectionHeader } from "../../components";
 import { convertToMoney } from "../../utils/string.util";
 import { listImgBase64 } from "../../resources/static/categoriesImages";
 import useTransaction from "../../hooks/useTransaction";
@@ -14,15 +15,27 @@ type TransactionsScreenProps = NativeStackScreenProps<RootStackParamList, 'Trans
 
 const TransactionsScreen: React.FC<TransactionsScreenProps> = ({ navigation }) => {
   const { 
+    filteredList, 
+    totalValue, 
+    searchValue, 
     listTransaction, 
+    setSearchValue,
     formatTransactionListToSection, 
     getBalanceFromTransactions 
   } = useTransaction();
 
+  const [showSearch, setShowSearch] = useState<boolean>(false);
   const [formatedTransactionList, setFormatedTransactionList] = useState<Array<{title: string, data: Array<Transaction>}>>();
 
   useEffect(() => {
     setFormatedTransactionList(formatTransactionListToSection(listTransaction));
+    navigation.setOptions({
+      headerRight: () => (
+        <TouchableOpacity onPress={() => setShowSearch((previousValue) => !previousValue)}>
+          <SearchIcon size='5' color='black' />
+        </TouchableOpacity>
+      )
+    })
   }, []);
 
   const renderSectionHeader = (title: string, value: number) => (
@@ -43,7 +56,18 @@ const TransactionsScreen: React.FC<TransactionsScreenProps> = ({ navigation }) =
   return (
     <>
       {
-        formatedTransactionList && formatedTransactionList.length > 0 && 
+        showSearch && 
+        <SearchInput
+          placeholder="Search transactions"
+          searchValue={searchValue}
+          listItems={filteredList}
+          totalValue={totalValue}
+          renderListItem={renderItem}
+          onChange={(text: string) => setSearchValue(text)}
+        />
+      }
+      {
+        (!filteredList || filteredList.length === 0) && formatedTransactionList && formatedTransactionList.length > 0 && 
         <SectionList
           contentContainerStyle={{paddingBottom: 80}}
           showsVerticalScrollIndicator={false}